feat: protect from multiple `MethodChannel.Result` invocation
