+++ android/src/main/java/io/ably/flutter/plugin/AblyMethodCallHandler.java
@@ -133,46 +133,13 @@ public class AblyMethodCallHandler implements MethodChannel.MethodCallHandler {
     _map.put(PlatformConstants.PlatformMethod.cryptoGenerateRandomKey, this::cryptoGenerateRandomKey);
   }
 
-  // MethodChannel.Result wrapper that responds on the platform thread.
-  //
-  // Plugins crash with "Methods marked with @UiThread must be executed on the main thread."
-  // This happens while making network calls in thread other than main thread
-  //
-  // https://github.com/flutter/flutter/issues/34993#issue-459900986
-  // https://github.com/aloisdeniel/flutter_geocoder/commit/bc34cfe473bfd1934fe098bb7053248b75200241
-  private static class MethodResultWrapper implements MethodChannel.Result {
-    private final MethodChannel.Result methodResult;
-    private final Handler handler;
-
-    MethodResultWrapper(MethodChannel.Result result) {
-      methodResult = result;
-      handler = new Handler(Looper.getMainLooper());
-    }
-
-    @Override
-    public void success(final Object result) {
-      handler.post(() -> methodResult.success(result));
-    }
-
-    @Override
-    public void error(
-        final String errorCode, final String errorMessage, final Object errorDetails) {
-      handler.post(() -> methodResult.error(errorCode, errorMessage, errorDetails));
-    }
-
-    @Override
-    public void notImplemented() {
-      handler.post(methodResult::notImplemented);
-    }
-  }
-
   private void handleAblyException(@NonNull MethodChannel.Result result, @NonNull AblyException e) {
     result.error(Integer.toString(e.errorInfo.code), e.getMessage(), e.errorInfo);
   }
 
   @Override
   public void onMethodCall(@NonNull MethodCall call, @NonNull MethodChannel.Result rawResult) {
-    final MethodChannel.Result result = new MethodResultWrapper(rawResult);
+    final MethodChannel.Result result = new SingleTimeMethodResult(new MainThreadMethodResult(rawResult));
     Log.v(TAG, String.format("onMethodCall: Ably Flutter platform method \"%s\" invoked.", call.method));
     final BiConsumer<MethodCall, MethodChannel.Result> handler = _map.get(call.method);
     if (null == handler) {
@@ -183,7 +150,7 @@ public class AblyMethodCallHandler implements MethodChannel.MethodCallHandler {
         // We have a handler for a method with this name so delegate to it.
         handler.accept(call, result);
       } catch (Exception e) {
-        Log.e(TAG, String.format("\"%s\" platform method received error during invocation", call.method), e);
+        Log.e(TAG, String.format("\"%s\" platform method received error during invocation, caused by: %s", call.method, e.getMessage()), e);
       }
     }
   }
